Refactor event registry and deserialization flow
Replaces EventNameRegistry with EventRegistry and moves it to the Persistent namespace. Updates EventRecord to use object for Payload and adds DeserializedEvent. EventHandlerRunner now relies on EventRecord for deserialized events. IEventStorage and related methods require EventRegistry for scheduled fetches, enabling storage-layer deserialization. Updates constructors and documentation to reflect these changes.
